refactor: Answer each library in the type it asked for
Squirrel takes a string length signed, OpenAL takes a buffer name signed though
it hands them out unsigned, and the scripting shims speak int on one side and
uint32_t on the other. Each of those crossings is now written where it happens
rather than left for the compiler to make quietly.

Two are corrections rather than casts. The ogg reader multiplied a sample count
by a size_t when it can hold the count's own type, and ReaderMapping's uint32_t
overload reads through as_int, which is the one place a level file can offer a
negative where an unsigned is wanted.

+++ src/audio/wav_sound_file.cpp
@@ -111,7 +111,7 @@ WavSoundFile::WavSoundFile(const std::string& filename) :
   if(encoding != 1)
     throw SoundError("only PCM encoding supported");
   channels = read16LE(file);
-  rate = read32LE(file);
+  rate = static_cast<int>(read32LE(file));
   uint32_t byterate = read32LE(file);
   (void) byterate;
   uint16_t blockalign = read16LE(file);
